R: consolidate MultipleArgumentsPerToken tests into a new file
