+++ src/System.CommandLine.Tests/OptionTests.cs
@@ -9,7 +9,7 @@ using Xunit;
 
 namespace System.CommandLine.Tests
 {
-    public class OptionTests : SymbolTests
+    public partial class OptionTests : SymbolTests
     {
         [Fact]
         public void When_an_option_has_only_one_alias_then_that_alias_is_its_name()
@@ -441,33 +441,6 @@ namespace System.CommandLine.Tests
             result.ValueForOption(option)
                 .Should()
                 .BeEmpty();
-        }
-
-        [Fact]
-        public void Single_option_arg_is_matched_when_disallowing_multiple_args_per_option_token()
-        {
-            var option = new Option<string[]>("--option") { AllowMultipleArgumentsPerToken = false };
-            var command = new Command("the-command") { option };
-
-            var result = command.Parse("--option 1 2");
-
-            var optionResult = result.ValueForOption(option);
-
-            optionResult.Should().BeEquivalentTo(new[] { "1" });
-            result.UnmatchedTokens.Should().BeEquivalentTo(new[] { "2" });
-        }
-
-        [Fact]
-        public void Multiple_option_args_are_matched_with_multiple_option_tokens_when_disallowing_multiple_args_per_option_token()
-        {
-            var option = new Option<string[]>("--option") { AllowMultipleArgumentsPerToken = false };
-            var command = new Command("the-command") { option };
-
-            var result = command.Parse("--option 1 --option 2");
-
-            var optionResult = result.ValueForOption(option);
-
-            optionResult.Should().BeEquivalentTo(new[] { "1", "2" });
         }
 
         [Fact]
