+++ src/System.CommandLine.Tests/ParserTests.cs
@@ -451,116 +451,6 @@ namespace System.CommandLine.Tests
         }
 
         [Fact]
-        public void When_a_Parser_root_option_is_not_respecified_but_limit_is_not_reached_then_the_following_token_is_used_as_value()
-        {
-            var animalsOption = new Option(new[] { "-a", "--animals" })
-            {
-                AllowMultipleArgumentsPerToken = true,
-                Arity = ArgumentArity.ZeroOrMore
-            };
-            var vegetablesOption = new Option(new[] { "-v", "--vegetables" }) { Arity = ArgumentArity.ZeroOrMore };
-
-            var parser = new Parser(
-                animalsOption,
-                vegetablesOption);
-
-            var result = parser.Parse("-a cat dog -v carrot");
-
-            result
-                .FindResultFor(animalsOption)
-                .Tokens
-                .Select(t => t.Value)
-                .Should()
-                .BeEquivalentTo(new[] { "cat", "dog" });
-
-            result
-                .FindResultFor(vegetablesOption)
-                .Tokens
-                .Select(t => t.Value)
-                .Should()
-                .BeEquivalentTo("carrot");
-
-            result
-                .UnmatchedTokens
-                .Should()
-                .BeNullOrEmpty();
-        }
-
-        [Fact]
-        public void When_a_Parser_root_option_is_not_respecified_and_limit_is_reached_then_the_following_token_is_unmatched()
-        {
-            var animalsOption = new Option(new[] { "-a", "--animals" })
-            {
-                AllowMultipleArgumentsPerToken = true,
-                Arity = ArgumentArity.ZeroOrOne
-            };
-            var vegetablesOption = new Option(new[] { "-v", "--vegetables" }) { Arity = ArgumentArity.ZeroOrMore };
-
-            var parser = new Parser(
-                animalsOption,
-                vegetablesOption);
-
-            ParseResult result = parser.Parse("-a cat some-arg -v carrot");
-
-            result.FindResultFor(animalsOption)
-                .Tokens
-                .Select(t => t.Value)
-                .Should()
-                .BeEquivalentTo("cat");
-
-            result.FindResultFor(vegetablesOption)
-                .Tokens
-                .Select(t => t.Value)
-                .Should()
-                .BeEquivalentTo("carrot");
-
-            result
-                .UnmatchedTokens
-                .Should()
-                .BeEquivalentTo("some-arg");
-        }
-
-        [Fact]
-        public void When_an_option_is_not_respecified_but_limit_is_not_reached_then_the_following_token_is_considered_as_value()
-        {
-            var animalsOption = new Option(new[] { "-a", "--animals" })
-            {
-                AllowMultipleArgumentsPerToken = true,
-                Arity = ArgumentArity.ZeroOrMore
-            };
-            var vegetablesOption = new Option(new[] { "-v", "--vegetables" }) { Arity = ArgumentArity.ZeroOrMore };
-            var parser = new Parser(
-                new Command("the-command")
-                {
-                    animalsOption,
-                    vegetablesOption
-                },
-                new Argument
-                {
-                    Arity = ArgumentArity.ZeroOrMore
-                });
-
-            var result = parser.Parse("the-command -a cat dog -v carrot");
-
-            result.FindResultFor(animalsOption)
-                  .Tokens
-                  .Select(t => t.Value)
-                  .Should()
-                  .BeEquivalentTo("cat", "dog");
-
-            result.FindResultFor(vegetablesOption)
-                  .Tokens
-                  .Select(t => t.Value)
-                  .Should()
-                  .BeEquivalentTo("carrot");
-
-            result.CommandResult
-                  .Tokens
-                  .Should()
-                  .BeNullOrEmpty();
-        }
-
-        [Fact]
         public void When_an_option_is_not_respecified_but_limit_is_reached_then_the_following_token_is_considered_an_argument_to_the_parent_command()
         {
             var animalsOption = new Option(new[] { "-a", "--animals" }) { Arity = ArgumentArity.ZeroOrOne };
